requirements.txt
# PySpark and Delta Lake
pyspark==3.4.1
delta-spark==2.4.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2

# Optional: For enhanced functionality
# pytest==7.4.3              # 테스트용
# jupyter==1.0.0             # Jupyter Notebook용
# matplotlib==3.8.2          # 시각화용
# seaborn==0.13.0            # 시각화용